Move GetProjectCallable to testutils
